+++ backend/src/main/java/shook/shook/voting_song/application/VotingSongPartService.java
@@ -1,18 +1,18 @@
-package shook.shook.song.application;
+package shook.shook.voting_song.application;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import shook.shook.part.domain.PartLength;
-import shook.shook.song.application.dto.voting_song.VotingSongPartRegisterRequest;
-import shook.shook.song.domain.repository.RegisterRepository;
-import shook.shook.song.domain.repository.VotingSongPartRepository;
-import shook.shook.song.domain.repository.VotingSongRepository;
-import shook.shook.song.domain.voting_song.Register;
-import shook.shook.song.domain.voting_song.VotingSong;
-import shook.shook.song.domain.voting_song.VotingSongPart;
-import shook.shook.song.exception.voting_song.VotingSongException.VotingSongNotExistException;
-import shook.shook.song.exception.voting_song.VotingSongPartException;
+import shook.shook.voting_song.application.dto.VotingSongPartRegisterRequest;
+import shook.shook.voting_song.domain.Register;
+import shook.shook.voting_song.domain.VotingSong;
+import shook.shook.voting_song.domain.VotingSongPart;
+import shook.shook.voting_song.domain.repository.RegisterRepository;
+import shook.shook.voting_song.domain.repository.VotingSongPartRepository;
+import shook.shook.voting_song.domain.repository.VotingSongRepository;
+import shook.shook.voting_song.exception.VotingSongException.VotingSongNotExistException;
+import shook.shook.voting_song.exception.VotingSongPartException;
 
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
