+++ backend/src/main/java/shook/shook/voting_song/application/VotingSongService.java
@@ -1,10 +1,10 @@
-package shook.shook.song.application;
+package shook.shook.voting_song.application;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import shook.shook.song.application.dto.voting_song.VotingSongRegisterRequest;
-import shook.shook.song.domain.repository.VotingSongRepository;
+import shook.shook.voting_song.application.dto.VotingSongRegisterRequest;
+import shook.shook.voting_song.domain.repository.VotingSongRepository;
 
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
