+++ backend/src/main/java/shook/shook/voting_song/application/dto/VotingSongRegisterRequest.java
@@ -1,4 +1,4 @@
-package shook.shook.song.application.dto.voting_song;
+package shook.shook.voting_song.application.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -7,7 +7,7 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import shook.shook.song.domain.voting_song.VotingSong;
+import shook.shook.voting_song.domain.VotingSong;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor
