+++ backend/src/main/java/shook/shook/voting_song/domain/repository/VotingSongPartRepository.java
@@ -1,10 +1,10 @@
-package shook.shook.song.domain.repository;
+package shook.shook.voting_song.domain.repository;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import shook.shook.song.domain.voting_song.VotingSong;
-import shook.shook.song.domain.voting_song.VotingSongPart;
+import shook.shook.voting_song.domain.VotingSong;
+import shook.shook.voting_song.domain.VotingSongPart;
 
 @Repository
 public interface VotingSongPartRepository extends JpaRepository<VotingSongPart, Long> {
