+++ backend/src/main/java/shook/shook/voting_song/domain/repository/VotingSongRepository.java
@@ -1,10 +1,10 @@
-package shook.shook.song.domain.repository;
+package shook.shook.voting_song.domain.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import shook.shook.song.domain.SongTitle;
-import shook.shook.song.domain.voting_song.VotingSong;
+import shook.shook.voting_song.domain.VotingSong;
 
 @Repository
 public interface VotingSongRepository extends JpaRepository<VotingSong, Long> {
