+++ backend/src/test/java/shook/shook/voting_song/domain/VotingSongPartTest.java
@@ -1,4 +1,4 @@
-package shook.shook.song.domain.voting_song;
+package shook.shook.voting_song.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import shook.shook.part.domain.PartLength;
 import shook.shook.part.exception.PartException;
-import shook.shook.song.exception.voting_song.RegisterException;
+import shook.shook.voting_song.exception.RegisterException;
 
 class VotingSongPartTest {
